requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.1.1
Werkzeug==0.11.4

